+++ backend/packages/provider-s3/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @cat-factory/provider-s3
 
+## 0.2.433
+
+### Patch Changes
+
+- Updated dependencies [ec0aba1]
+  - @cat-factory/kernel@0.328.0
+
 ## 0.2.432
 
 ### Patch Changes